pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "python-pptx>=1.0.2",
    "replit>=4.1.2",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
]
